+++ memberaudit/tests/models/test_characters_2.py
@@ -1,4 +1,3 @@
-import datetime as dt
 from unittest.mock import MagicMock, patch
 
 from django.test import TestCase
@@ -752,55 +751,6 @@ class TestCharacterResetTokenErrorNotifiedIfStatusOk(TestCase):
         # then
         character.refresh_from_db()
         self.assertIsNone(character.token_error_notified_at)
-
-
-class TestCharacterIsUpdateNeeded(TestCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        load_entities()
-        cls.character = create_memberaudit_character(1001)
-
-    @patch(MODULE_PATH + ".MEMBERAUDIT_FEATURE_ROLES_ENABLED", True)
-    def test_should_return_false_when_all_sections_are_current(self):
-        # given
-        for section in Character.UpdateSection:
-            create_character_update_status(self.character, section=section)
-
-        # when/then
-        self.assertFalse(self.character.is_update_needed())
-
-    @patch(MODULE_PATH + ".MEMBERAUDIT_FEATURE_ROLES_ENABLED", True)
-    def test_should_return_true_when_one_section_is_outdated(self):
-        # given
-        current_sections = set(Character.UpdateSection) - {
-            Character.UpdateSection.ASSETS
-        }
-        for section in current_sections:
-            create_character_update_status(self.character, section=section)
-
-        create_character_update_status(
-            self.character,
-            section=Character.UpdateSection.ASSETS,
-            started_at=now() - dt.timedelta(hours=24),
-        )
-
-        # when/then
-        self.assertTrue(self.character.is_update_needed())
-
-    @patch(MODULE_PATH + ".MEMBERAUDIT_FEATURE_ROLES_ENABLED", False)
-    def test_should_return_false_when_all_enabled_sections_are_current(self):
-        # given
-        for section in Character.UpdateSection.enabled_sections():
-            create_character_update_status(self.character, section=section)
-
-        create_character_update_status(
-            self.character,
-            section=Character.UpdateSection.ROLES,
-            started_at=now() - dt.timedelta(hours=24),
-        )
-        # when/then
-        self.assertFalse(self.character.is_update_needed())
 
 
 class TestCharacterGetEsiScopes(TestCase):